# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "All APIs related to Karchu.",
        "title": "Karchu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "karchu.onrender.com",
    "basePath": "/",
    "paths": {
        "/categories/all": {
            "post": {
                "description": "returns array of categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories of user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/categories/": {
            "post": {
                "description": "creates a category for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "create a category",
                "parameters": [
                    {
                        "description": "enter credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a category for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "description": "enter credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            },
            "patch": {
                "description": "Renames a category for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Rename a category",
                "parameters": [
                    {
                        "description": "enter credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RenameCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v1/categories/merge": {
            "post": {
                "description": "Merges a category into another, all the transactions of soruce category will now be part of destination category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Merge a category into another",
                "parameters": [
                    {
                        "description": "enter credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.MergeCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v1/categories/n": {
            "post": {
                "description": "returns array of categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories of user",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/transactions/categories": {
            "post": {
                "description": "returns transactions list for a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "returns transactions of category",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TransactionsOfCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/views.TransactionWithCategory"
                            }
                        }
                    }
                }
            }
        },
        "/v1/transactions/category": {
            "post": {
                "description": "returns transactions list for a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "returns transactions of category",
                "parameters": [
                    {
                        "description": "enter Email,Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TransactionsOfCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v2/friends/": {
            "post": {
                "description": "creates a friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "create a friend",
                "parameters": [
                    {
                        "description": "enter Email, Password and friend Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateFriendReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/friends/get": {
            "post": {
                "description": "get friends",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "get friends",
                "parameters": [
                    {
                        "description": "enter Email, Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetFriendsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requests.CreateCategoryReq": {
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "requests.CreateFriendReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "friendName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.DeleteCategoryReq": {
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.GetFriendsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.MergeCategory": {
            "type": "object",
            "properties": {
                "destinationCategoryName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "sourceCategoryName": {
                    "type": "string"
                }
            }
        },
...
        "requests.RenameCategoryReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "newCategoryName": {
                    "type": "string"
                },
                "oldCategoryName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.TransactionsOfCategoryReq": {
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.UserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "responses.SuccessRes": {
            "type": "object",
            "properties": {
                "success_code": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
        "views.TransactionWithCategory": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "splitTag": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}